+++ packages/choice-input/src/ChoiceGroupMixin.js
@@ -5,12 +5,27 @@ export const ChoiceGroupMixin = dedupeMixin(
   superclass =>
     // eslint-disable-next-line
     class ChoiceGroupMixin extends FormRegistrarMixin(InteractionStateMixin(superclass)) {
+      static get properties() {
+        return {
+          /**
+           * @desc When false (default), modelValue and serializedValue will reflect the
+           * currently selected choice (usually a string). When true, modelValue will and
+           * serializedValue will be an array of strings.
+           * @type {boolean}
+           */
+          multipleChoice: {
+            type: Boolean,
+            attribute: 'multiple-choice',
+          },
+        };
+      }
+
       get modelValue() {
         const elems = this._getCheckedElements();
         if (this.multipleChoice) {
           return elems.map(el => el.modelValue.value);
         }
-        return elems ? elems.modelValue.value : '';
+        return elems[0] ? elems[0].modelValue.value : '';
       }
 
       set modelValue(value) {
@@ -18,11 +33,19 @@ export const ChoiceGroupMixin = dedupeMixin(
       }
 
       get serializedValue() {
+        // We want to filter out disabled values out by default:
+        // The goal of serializing values could either be submitting state to a backend
+        // ot storing state in a backend. For this, only values that are entered by the end
+        // user are relevant, choice values are always defined by the Application Developer
+        // and known by the backend.
+
+        // Assuming values are always defined as strings, modelValues and serializedValues
+        // are the same.
         const elems = this._getCheckedElements();
         if (this.multipleChoice) {
-          return this.modelValue;
+          return elems.map(el => el.serializedValue.value);
         }
-        return elems ? elems.serializedValue : '';
+        return elems[0] ? elems[0].serializedValue.value : '';
       }
 
       set serializedValue(value) {
@@ -58,19 +81,15 @@ export const ChoiceGroupMixin = dedupeMixin(
       }
 
       /**
-       * @override from LionFieldset
+       * @override
        */
-      // eslint-disable-next-line class-methods-use-this
-      get _childrenCanHaveSameName() {
-        return true;
-      }
-
-      /**
-       * @override from LionFieldset
-       */
-      // eslint-disable-next-line class-methods-use-this
-      get _childNamesCanBeDuplicate() {
-        return true;
+      _getFromAllFormElements(property, filterCondition = () => true) {
+        // For modelValue and serializedValue, an exception should be made,
+        // The reset can be requested from children
+        if (property === 'modelValue' || property === 'serializedValue') {
+          return this[property];
+        }
+        return this.formElements.filter(filterCondition).map(el => el.property);
       }
 
       _throwWhenInvalidChildModelValue(child) {
@@ -108,7 +127,7 @@ export const ChoiceGroupMixin = dedupeMixin(
         if (target.checked === false) return;
 
         const groupName = target.name;
-        this.formElementsArray
+        this.formElements
           .filter(i => i.name === groupName)
           .forEach(choice => {
             if (choice !== target) {
@@ -119,11 +138,8 @@ export const ChoiceGroupMixin = dedupeMixin(
       }
 
       _getCheckedElements() {
-        const filtered = this.formElementsArray.filter(el => el.checked === true);
-        if (this.multipleChoice) {
-          return filtered;
-        }
-        return filtered.length > 0 ? filtered[0] : undefined;
+        // We want to filter out disabled values out by default
+        return this.formElements.filter(el => el.checked && !el.disabled);
       }
 
       async _setCheckedElements(value, check) {
@@ -131,12 +147,12 @@ export const ChoiceGroupMixin = dedupeMixin(
           await this.registrationReady;
         }
 
-        for (let i = 0; i < this.formElementsArray.length; i += 1) {
+        for (let i = 0; i < this.formElements.length; i += 1) {
           if (this.multipleChoice) {
-            this.formElementsArray[i].checked = value.includes(this.formElementsArray[i].value);
-          } else if (check(this.formElementsArray[i], value)) {
+            this.formElements[i].checked = value.includes(this.formElements[i].value);
+          } else if (check(this.formElements[i], value)) {
             // Allows checking against custom values e.g. formattedValue or serializedValue
-            this.formElementsArray[i].checked = true;
+            this.formElements[i].checked = true;
           }
         }
       }