+++ packages/field/index.js
@@ -1,9 +1,9 @@
-export { FieldCustomMixin } from './src/FieldCustomMixin.js';
 export { FocusMixin } from './src/FocusMixin.js';
 export { FormatMixin } from './src/FormatMixin.js';
 export { FormControlMixin } from './src/FormControlMixin.js';
 export { InteractionStateMixin } from './src/InteractionStateMixin.js'; // applies FocusMixin
 export { LionField } from './src/LionField.js';
-export { FormRegisteringMixin } from './src/FormRegisteringMixin.js';
-export { FormRegistrarMixin } from './src/FormRegistrarMixin.js';
-export { FormRegistrarPortalMixin } from './src/FormRegistrarPortalMixin.js';
+export { FormRegisteringMixin } from './src/registration/FormRegisteringMixin.js';
+export { FormRegistrarMixin } from './src/registration/FormRegistrarMixin.js';
+export { FormRegistrarPortalMixin } from './src/registration/FormRegistrarPortalMixin.js';
+export { FormControlsCollection } from './src/registration/FormControlsCollection.js';