+++ packages/field/src/LionField.js
@@ -37,10 +37,6 @@ export class LionField extends FormControlMixin(
         // make sure validation can be triggered based on observer
         type: Boolean,
       },
-      name: {
-        type: String,
-        reflect: true,
-      },
       autocomplete: {
         type: String,
         reflect: true,
@@ -216,13 +212,5 @@ export class LionField extends FormControlMixin(
     } else {
       this._inputNode.value = newValue;
     }
-  }
-
-  set fieldName(value) {
-    this.__fieldName = value;
-  }
-
-  get fieldName() {
-    return this.__fieldName || this.label || this.name;
   }
 }