+++ packages/fieldset/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.1](https://github.com/ing-bank/lion/compare/@lion/fieldset@0.8.0...@lion/fieldset@0.8.1) (2020-02-26)
+
+**Note:** Version bump only for package @lion/fieldset
+
+
+
+
+
+# [0.8.0](https://github.com/ing-bank/lion/compare/@lion/fieldset@0.7.2...@lion/fieldset@0.8.0) (2020-02-20)
+
+
+### Bug Fixes
+
+* rearrange/cleanup mixin duties ([cb7120e](https://github.com/ing-bank/lion/commit/cb7120e3a59f02e7e87f00c4d281e47ca8d33ea8))
+
+
+### Features
+
+* api normalisation formElements and values ([9b905c4](https://github.com/ing-bank/lion/commit/9b905c492a0c0d2226cc1d75c73e2e70dc97815a))
+
+
+
+
+
 ## [0.7.2](https://github.com/ing-bank/lion/compare/@lion/fieldset@0.7.1...@lion/fieldset@0.7.2) (2020-02-19)