+++ packages/form-system/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/ing-bank/lion/compare/@lion/form-system@0.5.0...@lion/form-system@0.5.1) (2020-02-26)
+
+**Note:** Version bump only for package @lion/form-system
+
+
+
+
+
+# [0.5.0](https://github.com/ing-bank/lion/compare/@lion/form-system@0.4.6...@lion/form-system@0.5.0) (2020-02-20)
+
+
+### Features
+
+* api normalisation formElements and values ([9b905c4](https://github.com/ing-bank/lion/commit/9b905c492a0c0d2226cc1d75c73e2e70dc97815a))
+
+
+
+
+
 ## [0.4.6](https://github.com/ing-bank/lion/compare/@lion/form-system@0.4.5...@lion/form-system@0.4.6) (2020-02-19)