+++ packages/select-rich/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.1](https://github.com/ing-bank/lion/compare/@lion/select-rich@0.11.0...@lion/select-rich@0.11.1) (2020-02-26)
+
+**Note:** Version bump only for package @lion/select-rich
+
+
+
+
+
+# [0.11.0](https://github.com/ing-bank/lion/compare/@lion/select-rich@0.10.4...@lion/select-rich@0.11.0) (2020-02-20)
+
+
+### Features
+
+* api normalisation formElements and values ([9b905c4](https://github.com/ing-bank/lion/commit/9b905c492a0c0d2226cc1d75c73e2e70dc97815a))
+
+
+
+
+
 ## [0.10.4](https://github.com/ing-bank/lion/compare/@lion/select-rich@0.10.3...@lion/select-rich@0.10.4) (2020-02-19)