+++ packages/textarea/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.15](https://github.com/ing-bank/lion/compare/@lion/textarea@0.5.14...@lion/textarea@0.5.15) (2020-02-26)
+
+**Note:** Version bump only for package @lion/textarea
+
+
+
+
+
+## [0.5.14](https://github.com/ing-bank/lion/compare/@lion/textarea@0.5.13...@lion/textarea@0.5.14) (2020-02-20)
+
+**Note:** Version bump only for package @lion/textarea
+
+
+
+
+
 ## [0.5.13](https://github.com/ing-bank/lion/compare/@lion/textarea@0.5.12...@lion/textarea@0.5.13) (2020-02-19)