+++ packages/validate/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.8](https://github.com/ing-bank/lion/compare/@lion/validate@0.6.7...@lion/validate@0.6.8) (2020-02-26)
+
+**Note:** Version bump only for package @lion/validate
+
+
+
+
+
 ## [0.6.7](https://github.com/ing-bank/lion/compare/@lion/validate@0.6.6...@lion/validate@0.6.7) (2020-02-19)